requirements.txt
 flask==2.1.0
flask-cors==4.0.0
pymongo==4.3.3
pinecone-client==2.2.4
torch==2.1.2
numpy==1.24.3
transformers==4.29.2
faiss-cpu==1.7.4
pytest==7.2.0
python-dotenv==0.21.0
PyPDF2==3.0.1
reportlab==4.2.5
Werkzeug==2.1.1
 flask==2.3.2
flask-cors==4.0.0
pymongo==4.6.1
pinecone-client==2.2.4
torch==2.1.2
numpy==1.24.3
transformers==4.36.2
faiss-cpu==1.7.4
pytest==7.4.4
python-dotenv==1.0.0
